abstract the code changes
