+++ accounts/abi/bind/backends/simulated_zkevm.go
@@ -1,0 +1,7 @@
+package backends
+
+import zktypes "github.com/ledgerwatch/erigon/zk/types"
+
+func (m callMsg) EffectiveGasPricePercentage() uint8 {
+	return zktypes.EFFECTIVE_GAS_PRICE_PERCENTAGE_DISABLED
+}
