+++ cmd/hack/tool/fromdb/tool.go
@@ -3,8 +3,8 @@ package fromdb
 import (
 	"context"
 
-	"github.com/ledgerwatch/erigon-lib/kv"
 	"github.com/ledgerwatch/erigon/chain"
+	"github.com/ledgerwatch/erigon-lib/kv"
 
 	"github.com/ledgerwatch/erigon/cmd/hack/tool"
 	"github.com/ledgerwatch/erigon/ethdb/prune"
