+++ cmd/hack/tool/tool.go
@@ -3,8 +3,8 @@ package tool
 import (
 	"strconv"
 
-	"github.com/ledgerwatch/erigon-lib/kv"
 	"github.com/ledgerwatch/erigon/chain"
+	"github.com/ledgerwatch/erigon-lib/kv"
 
 	"github.com/ledgerwatch/erigon/core/rawdb"
 )
