+++ cmd/rpcdaemon/rpcservices/eth_backend.go
@@ -317,7 +317,3 @@ func (back *RemoteBackend) PendingBlock(ctx context.Context) (*types.Block, erro
 
 	return &block, nil
 }
-
-func (back *RemoteBackend) TxnEffectiveGasPricePercentage(ctx context.Context, tx kv.Tx, txnHash libcommon.Hash) (uint8, error) {
-	return back.blockReader.TxnEffectiveGasPricePercentage(ctx, tx, txnHash)
-}
