+++ core/genesis_write.go
@@ -49,7 +49,6 @@ import (
 	"github.com/ledgerwatch/erigon/params/networkname"
 	"github.com/ledgerwatch/erigon/smt/pkg/db"
 	"github.com/ledgerwatch/erigon/smt/pkg/smt"
-	"github.com/ledgerwatch/erigon/zkevm/hex"
 )
 
 // CommitGenesisBlock writes or updates the genesis block in db.
@@ -439,76 +438,6 @@ func ChiadoGenesisBlock() *types.Genesis {
 	}
 }
 
-func HermezMainnetGenesisBlock() *types.Genesis {
-	return &types.Genesis{
-		Config:     params.HermezMainnetChainConfig,
-		Timestamp:  1679653163,
-		GasLimit:   0x0,
-		Difficulty: big.NewInt(0x0),
-		Alloc:      readPrealloc("allocs/hermez.json"),
-	}
-}
-
-func HermezTestnetGenesisBlock() *types.Genesis {
-	return &types.Genesis{
-		Config:     params.HermezTestnetChainConfig,
-		Timestamp:  1677601932,
-		GasLimit:   0x0,
-		Difficulty: big.NewInt(0x0),
-		Alloc:      readPrealloc("allocs/hermez-testnet.json"),
-	}
-}
-
-func HermezBlueberryGenesisBlock() *types.Genesis {
-	return &types.Genesis{
-		Config:     params.HermezBlueberryChainConfig,
-		Timestamp:  1676996964,
-		GasLimit:   0x0,
-		Difficulty: big.NewInt(0x0),
-		Alloc:      readPrealloc("allocs/hermez-blueberry.json"),
-	}
-}
-
-func HermezCardonaGenesisBlock() *types.Genesis {
-	return &types.Genesis{
-		Config:     params.HermezCardonaChainConfig,
-		Timestamp:  1676996964,
-		GasLimit:   0x0,
-		Difficulty: big.NewInt(0x0),
-		Alloc:      readPrealloc("allocs/hermez-cardona.json"),
-	}
-}
-
-func HermezCardonaInternalGenesisBlock() *types.Genesis {
-	return &types.Genesis{
-		Config:     params.HermezCardonaInternalChainConfig,
-		Timestamp:  1676996964,
-		GasLimit:   0x0,
-		Difficulty: big.NewInt(0x0),
-		Alloc:      readPrealloc("allocs/hermez-cardona-internal.json"),
-	}
-}
-
-func HermezLocalDevnetGenesisBlock() *types.Genesis {
-	return &types.Genesis{
-		Config:     params.HermezLocalDevnetChainConfig,
-		Timestamp:  1676996964,
-		GasLimit:   0x0,
-		Difficulty: big.NewInt(0x0),
-		Alloc:      readPrealloc("allocs/hermez-dev.json"),
-	}
-}
-
-func X1TestnetGenesisBlock() *types.Genesis {
-	return &types.Genesis{
-		Config:     params.X1TestnetChainConfig,
-		Timestamp:  1699369668,
-		GasLimit:   0x0,
-		Difficulty: big.NewInt(0x0),
-		Alloc:      readPrealloc("allocs/x1-testnet.json"),
-	}
-}
-
 // Pre-calculated version of:
 //
 //	DevnetSignPrivateKey = crypto.HexToECDSA(sha256.Sum256([]byte("erigon devnet key")))
@@ -534,39 +463,6 @@ func DeveloperGenesisBlock(period uint64, faucet libcommon.Address) *types.Genes
 
 var genesisTmpDB kv.RwDB
 var genesisDBLock sync.Mutex
-
-func processAccount(s *smt.SMT, root *big.Int, a *types.GenesisAccount, addr libcommon.Address) (*big.Int, error) {
-
-	if root == nil {
-		root = big.NewInt(0)
-	}
-
-	// store the account balance and nonce
-	r, err := s.SetAccountState(addr.String(), a.Balance, new(big.Int).SetUint64(a.Nonce))
-	if err != nil {
-		return nil, err
-	}
-
-	if len(a.Code) > 0 {
-		xs := hex.EncodeToString(a.Code)
-		err = s.SetContractBytecode(addr.String(), xs)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	// parse the storage into map[string]string by splitting the storage hex into two 32 bit values
-	sm := make(map[string]string)
-	for k, v := range a.Storage {
-		sm[k.String()] = v.String()
-	}
-
-	// store the account storage
-	if len(sm) > 0 {
-		r, err = s.SetContractStorage(addr.String(), sm)
-	}
-	return r, nil
-}
 
 // ToBlock creates the genesis block and writes state of a genesis specification
 // to the given database (or discards it if nil).
