+++ core/rawdb/accessors_chain.go
@@ -41,9 +41,9 @@ import (
 	"github.com/ledgerwatch/erigon/common"
 	"github.com/ledgerwatch/erigon/common/dbutils"
 	"github.com/ledgerwatch/erigon/core/types"
+	"github.com/ledgerwatch/erigon/eth/stagedsync/stages"
 	"github.com/ledgerwatch/erigon/ethdb/cbor"
 	"github.com/ledgerwatch/erigon/rlp"
-	"github.com/ledgerwatch/erigon/eth/stagedsync/stages"
 	"github.com/ledgerwatch/erigon/turbo/services"
 )
 
@@ -493,27 +493,6 @@ func WriteTransactions(db kv.RwTx, txs []types.Transaction, baseTxId uint64, blo
 	return nil
 }
 
-func DeleteTransactions(db kv.RwTx, txsCount, baseTxId uint64, blockHash *libcommon.Hash) error {
-	for id := baseTxId; id < baseTxId+txsCount; id++ {
-		txIdKey := make([]byte, 8)
-		binary.BigEndian.PutUint64(txIdKey, id)
-
-		var err error
-		if blockHash != nil {
-			key := append(txIdKey, blockHash.Bytes()...)
-			db.Delete(kv.EthTxV3, key)
-		} else {
-			db.Delete(kv.EthTx, txIdKey)
-		}
-
-		if err != nil {
-			return fmt.Errorf("error deleting tx: %w", err)
-		}
-	}
-
-	return nil
-}
-
 func WriteRawTransactions(tx kv.RwTx, txs [][]byte, baseTxId uint64, blockHash *common2.Hash) error {
 	txId := baseTxId
 	for _, txn := range txs {
@@ -750,40 +729,6 @@ func WriteBody(db kv.RwTx, hash libcommon.Hash, number uint64, body *types.Body)
 	}
 	if err != nil {
 		return fmt.Errorf("failed to WriteTransactions: %w", err)
-	}
-	return nil
-}
-
-func TruncateBodies(tx kv.RwTx, blockNum uint64) error {
-	if err := tx.ForEach(kv.BlockBody, hexutility.EncodeTs(blockNum), func(k, v []byte) error {
-		var body types.BodyForStorage
-		if err := rlp.DecodeBytes(v, &body); err != nil {
-			return fmt.Errorf("failed to decode body: %w", err)
-		}
-
-		txs, err := CanonicalTransactions(tx, body.BaseTxId, body.TxAmount)
-		if err != nil {
-			return fmt.Errorf("failed to read txs: %w", err)
-		}
-
-		blockhash := libcommon.BytesToHash(k)
-		// delete body for storage
-		deleteBody(tx, blockhash, blockNum)
-
-		// TODO: decrement sequence?
-		// decrement txs sequence
-		// if err := tx.DecrementSequence(kv.EthTx, uint64(body.TxAmount)); err != nil {
-		// 	return fmt.Errorf("failed to decrement sequence: %w", err)
-		// }
-
-		// delete transactions
-		if err := DeleteTransactions(tx, uint64(len(txs)), body.BaseTxId, &blockhash); err != nil {
-			return fmt.Errorf("failed to delete txs: %w", err)
-		}
-
-		return nil
-	}); err != nil {
-		return fmt.Errorf("TruncateBodies: %w", err)
 	}
 	return nil
 }
