+++ core/state/intra_block_state.go
@@ -18,28 +18,20 @@
 package state
 
 import (
-	"errors"
 	"fmt"
 	"sort"
 
 	"encoding/hex"
 	"github.com/holiman/uint256"
-	"github.com/iden3/go-iden3-crypto/keccak256"
 	libcommon "github.com/ledgerwatch/erigon-lib/common"
 	types2 "github.com/ledgerwatch/erigon-lib/types"
 	"github.com/ledgerwatch/erigon/chain"
-	"github.com/ledgerwatch/erigon/common"
 	"github.com/ledgerwatch/erigon/common/u256"
 	"github.com/ledgerwatch/erigon/core/types"
 	"github.com/ledgerwatch/erigon/core/types/accounts"
 	"github.com/ledgerwatch/erigon/smt/pkg/utils"
 	"github.com/ledgerwatch/erigon/turbo/trie"
 )
-
-type ReadOnlyHermezDb interface {
-	GetEffectiveGasPricePercentage(txHash libcommon.Hash) (uint8, error)
-	GetStateRoot(l2BlockNo uint64) (libcommon.Hash, error)
-}
 
 type revision struct {
 	id           int
@@ -249,14 +241,6 @@ func (sdb *IntraBlockState) GetCodeSize(addr libcommon.Address) int {
 		sdb.setErrorUnsafe(err)
 	}
 	return l
-}
-
-func (sdb *IntraBlockState) GetTxCount() (uint64, error) {
-	counter, ok := sdb.stateReader.(TxCountReader)
-	if !ok {
-		return 0, errors.New("state reader does not support GetTxCount")
-	}
-	return counter.GetTxCount()
 }
 
 // DESCRIBED: docs/programmers_guide/guide.md#address---identifier-of-an-account
@@ -800,49 +784,4 @@ func (sdb *IntraBlockState) AddressInAccessList(addr libcommon.Address) bool {
 // SlotInAccessList returns true if the given (address, slot)-tuple is in the access list.
 func (sdb *IntraBlockState) SlotInAccessList(addr libcommon.Address, slot libcommon.Hash) (addressPresent bool, slotPresent bool) {
 	return sdb.accessList.Contains(addr, slot)
-}
-
-func (sdb *IntraBlockState) ScalableSetTxNum() {
-	saddr := libcommon.HexToAddress("0x000000000000000000000000000000005ca1ab1e")
-	sl0 := libcommon.HexToHash("0x0")
-
-	txNum := uint256.NewInt(0)
-	sdb.GetState(saddr, &sl0, txNum)
-
-	txNum.Add(txNum, uint256.NewInt(1))
-
-	if !sdb.Exist(saddr) {
-		// create account if not exists
-		sdb.CreateAccount(saddr, true)
-	}
-
-	// set incremented tx num in state
-	sdb.SetState(saddr, &sl0, *txNum)
-}
-
-func (sdb *IntraBlockState) ScalableSetSmtRootHash(roHermezDb ReadOnlyHermezDb) error {
-	saddr := libcommon.HexToAddress("0x000000000000000000000000000000005ca1ab1e")
-	sl0 := libcommon.HexToHash("0x0")
-
-	txNum := uint256.NewInt(0)
-	sdb.GetState(saddr, &sl0, txNum)
-
-	// create mapping with keccak256(txnum,1) -> smt root
-	d1 := common.LeftPadBytes(txNum.Bytes(), 32)
-	d2 := common.LeftPadBytes(uint256.NewInt(1).Bytes(), 32)
-	mapKey := keccak256.Hash(d1, d2)
-	mkh := libcommon.BytesToHash(mapKey)
-
-	rpcHash, err := roHermezDb.GetStateRoot(txNum.Uint64())
-	if err != nil {
-		return err
-	}
-
-	if txNum.Uint64() >= 1 {
-		// set mapping of keccak256(txnum,1) -> smt root
-		rpcHashU256 := uint256.NewInt(0).SetBytes(rpcHash.Bytes())
-		sdb.SetState(saddr, &mkh, *rpcHashU256)
-	}
-
-	return nil
 }
