+++ core/vm/instructions_test.go
@@ -21,16 +21,12 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"math/big"
 	"os"
 	"testing"
-
-	types2 "github.com/ledgerwatch/erigon-lib/types"
 
 	"github.com/holiman/uint256"
 	libcommon "github.com/ledgerwatch/erigon-lib/common"
 
-	"github.com/ledgerwatch/erigon/core/types"
 	"github.com/ledgerwatch/erigon/core/vm/evmtypes"
 
 	"github.com/ledgerwatch/erigon/common"
@@ -230,111 +226,6 @@ func TestAddMod(t *testing.T) {
 		stack.Push(y)
 		stack.Push(x)
 		opAddmod(&pc, evmInterpreter, &ScopeContext{nil, stack, nil})
-		actual := stack.Pop()
-		if actual.Cmp(expected) != 0 {
-			t.Errorf("Testcase %d, expected  %x, got %x", i, expected, actual)
-		}
-	}
-}
-func TestBlockhashV2(t *testing.T) {
-
-	gethashFn := func(bn uint64) libcommon.Hash {
-		return libcommon.BigToHash(new(big.Int).SetUint64(bn))
-	}
-
-	var (
-		env            = NewEVM(evmtypes.BlockContext{GetHash: gethashFn}, evmtypes.TxContext{}, TestIntraBlockState{}, params.TestChainConfig, Config{})
-		stack          = stack.New()
-		evmInterpreter = NewZKEVMInterpreter(env, env.Config())
-		pc             = uint64(0)
-	)
-
-	tests := []struct {
-		blockNumber uint64
-		expected    libcommon.Hash
-	}{
-		{
-			blockNumber: 2,
-			expected:    libcommon.BigToHash(new(big.Int).SetUint64(2)),
-		}, {
-			blockNumber: 2000,
-			expected:    libcommon.BigToHash(new(big.Int).SetUint64(2000)),
-		}, {
-			blockNumber: 2000000000,
-			expected:    libcommon.BigToHash(new(big.Int).SetUint64(2000000000)),
-		},
-	}
-
-	for i, test := range tests {
-		expected := new(uint256.Int).SetBytes(test.expected.Bytes())
-		blockNumberBytes := new(uint256.Int).SetUint64(test.blockNumber)
-		stack.Push(blockNumberBytes)
-
-		opBlockhash_zkevm(&pc, evmInterpreter, &ScopeContext{nil, stack, nil})
-		actual := stack.Pop()
-
-		fmt.Println(actual)
-
-		if actual.Cmp(expected) != 0 {
-			t.Errorf("Testcase %d, expected  %x, got %x", i, expected, actual)
-		}
-	}
-}
-
-func TestDifficultyV2(t *testing.T) {
-	var (
-		env            = NewEVM(evmtypes.BlockContext{}, evmtypes.TxContext{}, nil, params.TestChainConfig, Config{})
-		stack          = stack.New()
-		evmInterpreter = NewEVMInterpreter(env, env.Config())
-		pc             = uint64(0)
-	)
-
-	zeroInt := new(big.Int).SetUint64(0)
-	v, _ := uint256.FromBig(zeroInt)
-
-	tests := []struct {
-		expected *uint256.Int
-	}{
-		{
-			expected: v,
-		},
-	}
-
-	for i, test := range tests {
-		expected := new(uint256.Int).SetBytes(test.expected.Bytes())
-
-		opDifficulty_zkevm(&pc, evmInterpreter, &ScopeContext{nil, stack, nil})
-		actual := stack.Pop()
-
-		if actual.Cmp(expected) != 0 {
-			t.Errorf("Testcase %d, expected  %x, got %x", i, expected, actual)
-		}
-	}
-}
-
-func TestExtCodeHashV2(t *testing.T) {
-	var (
-		env            = NewEVM(evmtypes.BlockContext{}, evmtypes.TxContext{}, TestIntraBlockState{}, params.TestChainConfig, Config{})
-		stack          = stack.New()
-		evmInterpreter = NewEVMInterpreter(env, env.Config())
-		pc             = uint64(0)
-	)
-	tests := []struct {
-		address  string
-		expected libcommon.Hash
-	}{
-		{"0000000000000000000000000000000000000000000000000000000000000000",
-			libcommon.HexToHash("0xtest"),
-		}, {"000000000000000000000000000000000000000000000000000000000fffffff",
-			libcommon.HexToHash("0xtest"),
-		},
-	}
-
-	for i, test := range tests {
-		address := new(uint256.Int).SetBytes(common.Hex2Bytes(test.address))
-		expected := new(uint256.Int).SetBytes(test.expected.Bytes())
-		stack.Push(address)
-		opExtCodeHash_zkevm(&pc, evmInterpreter, &ScopeContext{nil, stack, nil})
 		actual := stack.Pop()
 		if actual.Cmp(expected) != 0 {
 			t.Errorf("Testcase %d, expected  %x, got %x", i, expected, actual)
@@ -762,42 +653,3 @@ func TestCreate2Addreses(t *testing.T) {
 		}
 	}
 }
-
-type TestIntraBlockState struct{}
-
-func (ibs TestIntraBlockState) CreateAccount(libcommon.Address, bool) {}
-func (ibs TestIntraBlockState) GetTxCount() (uint64, error)           { return 0, nil }
-
-func (ibs TestIntraBlockState) SubBalance(libcommon.Address, *uint256.Int) {}
-func (ibs TestIntraBlockState) AddBalance(libcommon.Address, *uint256.Int) {}
-func (ibs TestIntraBlockState) GetBalance(libcommon.Address) *uint256.Int  { return nil }
-func (ibs TestIntraBlockState) GetNonce(libcommon.Address) uint64          { return 0 }
-func (ibs TestIntraBlockState) SetNonce(libcommon.Address, uint64)         {}
-func (ibs TestIntraBlockState) GetCodeHash(libcommon.Address) libcommon.Hash {
-	return libcommon.HexToHash("0xtest")
-}
-func (ibs TestIntraBlockState) GetCode(libcommon.Address) []byte                                   { return nil }
-func (ibs TestIntraBlockState) SetCode(libcommon.Address, []byte)                                  {}
-func (ibs TestIntraBlockState) GetCodeSize(libcommon.Address) int                                  { return 0 }
-func (ibs TestIntraBlockState) AddRefund(uint64)                                                   {}
-func (ibs TestIntraBlockState) SubRefund(uint64)                                                   {}
-func (ibs TestIntraBlockState) GetRefund() uint64                                                  { return 0 }
-func (ibs TestIntraBlockState) GetCommittedState(libcommon.Address, *libcommon.Hash, *uint256.Int) {}
-func (ibs TestIntraBlockState) GetState(address libcommon.Address, slot *libcommon.Hash, outValue *uint256.Int) {
-}
-func (ibs TestIntraBlockState) SetState(libcommon.Address, *libcommon.Hash, uint256.Int) {}
-func (ibs TestIntraBlockState) Selfdestruct(libcommon.Address) bool                      { return false }
-func (ibs TestIntraBlockState) HasSelfdestructed(libcommon.Address) bool                 { return false }
-func (ibs TestIntraBlockState) Exist(libcommon.Address) bool                             { return false }
-func (ibs TestIntraBlockState) Empty(libcommon.Address) bool                             { return false }
-func (ibs TestIntraBlockState) PrepareAccessList(sender libcommon.Address, dest *libcommon.Address, precompiles []libcommon.Address, txAccesses types2.AccessList) {
-}
-func (ibs TestIntraBlockState) AddressInAccessList(addr libcommon.Address) bool { return false }
-func (ibs TestIntraBlockState) SlotInAccessList(addr libcommon.Address, slot libcommon.Hash) (addressOk bool, slotOk bool) {
-	return false, false
-}
-func (ibs TestIntraBlockState) AddAddressToAccessList(addr libcommon.Address)                   {}
-func (ibs TestIntraBlockState) AddSlotToAccessList(addr libcommon.Address, slot libcommon.Hash) {}
-func (ibs TestIntraBlockState) RevertToSnapshot(int)                                            {}
-func (ibs TestIntraBlockState) Snapshot() int                                                   { return 0 }
-func (ibs TestIntraBlockState) AddLog(*types.Log)                                               {}
