+++ eth/stagedsync/stage_hashstate_test.go
@@ -11,7 +11,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	libcommon "github.com/ledgerwatch/erigon-lib/common"
-	"github.com/ledgerwatch/erigon/eth/stagedsync/stages"
 )
 
 func TestPromoteHashedStateClearState(t *testing.T) {
@@ -87,8 +86,8 @@ func TestUnwindHashed(t *testing.T) {
 	if err != nil {
 		t.Errorf("error while promoting state: %v", err)
 	}
-	u := &sync.UnwindState{UnwindPoint: 50}
-	s := &sync.StageState{BlockNumber: 100}
+	u := &UnwindState{UnwindPoint: 50}
+	s := &StageState{BlockNumber: 100}
 	err = unwindHashStateStageImpl("logPrefix", u, s, tx2, StageHashStateCfg(db2, dirs, historyV3, nil), context.Background())
 	if err != nil {
 		t.Errorf("error while unwind state: %v", err)
@@ -197,8 +196,8 @@ func TestUnwindHashStateShutdown(t *testing.T) {
 				require.NoError(t, err)
 			}
 
-			u := &sync.UnwindState{UnwindPoint: 5}
-			s := &sync.StageState{BlockNumber: 10}
+			u := &UnwindState{UnwindPoint: 5}
+			s := &StageState{BlockNumber: 10}
 			if err = unwindHashStateStageImpl("logPrefix", u, s, tx, cfg, ctx); !errors.Is(err, tc.errExp) {
 				t.Errorf("error does not match expected error while shutdown unwindHashStateStageImpl, got: %v, expected: %v", err, tc.errExp)
 			}
