+++ eth/stagedsync/stagebuilder.go
@@ -31,10 +31,8 @@ func MiningStages(
 			Forward: func(firstCycle bool, badBlockUnwind bool, s *StageState, u Unwinder, tx kv.RwTx, quiet bool) error {
 				return SpawnMiningCreateBlockStage(s, tx, createBlockCfg, ctx.Done())
 			},
-			Unwind: func(firstCycle bool, u *UnwindState, s *StageState, tx kv.RwTx) error {
-				return nil
-			},
-			Prune: func(firstCycle bool, u *PruneState, tx kv.RwTx) error { return nil },
+			Unwind: func(firstCycle bool, u *UnwindState, s *StageState, tx kv.RwTx) error { return nil },
+			Prune:  func(firstCycle bool, u *PruneState, tx kv.RwTx) error { return nil },
 		},
 		{
 			ID:          stages.MiningExecution,
@@ -42,10 +40,8 @@ func MiningStages(
 			Forward: func(firstCycle bool, badBlockUnwind bool, s *StageState, u Unwinder, tx kv.RwTx, quiet bool) error {
 				return SpawnMiningExecStage(s, tx, execCfg, ctx.Done())
 			},
-			Unwind: func(firstCycle bool, u *UnwindState, s *StageState, tx kv.RwTx) error {
-				return nil
-			},
-			Prune: func(firstCycle bool, u *PruneState, tx kv.RwTx) error { return nil },
+			Unwind: func(firstCycle bool, u *UnwindState, s *StageState, tx kv.RwTx) error { return nil },
+			Prune:  func(firstCycle bool, u *PruneState, tx kv.RwTx) error { return nil },
 		},
 		{
 			ID:          stages.HashState,
@@ -53,10 +49,8 @@ func MiningStages(
 			Forward: func(firstCycle bool, badBlockUnwind bool, s *StageState, u Unwinder, tx kv.RwTx, quiet bool) error {
 				return SpawnHashStateStage(s, tx, hashStateCfg, ctx, quiet)
 			},
-			Unwind: func(firstCycle bool, u *UnwindState, s *StageState, tx kv.RwTx) error {
-				return nil
-			},
-			Prune: func(firstCycle bool, u *PruneState, tx kv.RwTx) error { return nil },
+			Unwind: func(firstCycle bool, u *UnwindState, s *StageState, tx kv.RwTx) error { return nil },
+			Prune:  func(firstCycle bool, u *PruneState, tx kv.RwTx) error { return nil },
 		},
 		{
 			ID:          stages.IntermediateHashes,
@@ -69,10 +63,8 @@ func MiningStages(
 				createBlockCfg.miner.MiningBlock.Header.Root = stateRoot
 				return nil
 			},
-			Unwind: func(firstCycle bool, u *UnwindState, s *StageState, tx kv.RwTx) error {
-				return nil
-			},
-			Prune: func(firstCycle bool, u *PruneState, tx kv.RwTx) error { return nil },
+			Unwind: func(firstCycle bool, u *UnwindState, s *StageState, tx kv.RwTx) error { return nil },
+			Prune:  func(firstCycle bool, u *PruneState, tx kv.RwTx) error { return nil },
 		},
 		{
 			ID:          stages.MiningFinish,
@@ -80,10 +72,8 @@ func MiningStages(
 			Forward: func(firstCycle bool, badBlockUnwind bool, s *StageState, u Unwinder, tx kv.RwTx, quiet bool) error {
 				return SpawnMiningFinishStage(s, tx, finish, ctx.Done())
 			},
-			Unwind: func(firstCycle bool, u *UnwindState, s *StageState, tx kv.RwTx) error {
-				return nil
-			},
-			Prune: func(firstCycle bool, u *PruneState, tx kv.RwTx) error { return nil },
+			Unwind: func(firstCycle bool, u *UnwindState, s *StageState, tx kv.RwTx) error { return nil },
+			Prune:  func(firstCycle bool, u *PruneState, tx kv.RwTx) error { return nil },
 		},
 	}
 }
